+++ src/Stub/Document.php
@@ -546,18 +546,18 @@ trait Document {
 	/**
 	 * @param string $source
 	 * @param int $options
-	 * @return \Wikimedia\IDLeDOM\Document|bool
+	 * @return bool
 	 */
-	public function loadHTML( string $source, int $options = 0 ) {
+	public function loadHTML( string $source, int $options = 0 ): bool {
 		throw self::_unimplemented();
 	}
 
 	/**
 	 * @param string $source
 	 * @param int $options
-	 * @return \Wikimedia\IDLeDOM\Document|bool
+	 * @return bool
 	 */
-	public function loadXML( string $source, int $options = 0 ) {
+	public function loadXML( string $source, int $options = 0 ): bool {
 		throw self::_unimplemented();
 	}